lib/screens/requirements.txt
flask==2.2.5
numpy==1.24.4
pillow==9.5.0
torch==1.13.1
torchvision==0.14.1
gdowngdown==4.5.1 